SMK-S3 synth stability and patch migration updates
This change hardens the ESP32-S3 SMK firmware build, improves ESP-IDF tool resolution in local and CI scripts, and adds guardrails for note-off, FM, and filter behavior. It also introduces storage base-path support, legacy v3/v4 patch migration to the v5 format, filter inheritance handling for AMY presets, and expanded host-side validation for chorus depth, FM relative controls, patch persistence, and CRC checks.

+++ smk-s3/components/midi/midi_learn.cpp
@@ -56,7 +56,7 @@ void MidiLearn::finishAndSave() {
     target_profile_->crc32 = calculateProfileCrc32(*target_profile_);
     current_step_ = LearnStep::Complete;
     ESP_LOGI(TAG, "==================================================");
-    ESP_LOGI(TAG, "MIDI Learn Finished and Saved (CRC32: 0x%08X)", target_profile_->crc32);
+    ESP_LOGI(TAG, "MIDI Learn Finished and Saved (CRC32: 0x%08X)", static_cast<unsigned int>(target_profile_->crc32));
     ESP_LOGI(TAG, "==================================================");
 }
 
@@ -181,7 +181,7 @@ void MidiLearn::advanceStep() {
     if (current_step_ == LearnStep::Complete && target_profile_) {
         target_profile_->crc32 = calculateProfileCrc32(*target_profile_);
         ESP_LOGI(TAG, "==================================================");
-        ESP_LOGI(TAG, "MIDI Learn Wizard Completed Successfully! (CRC32: 0x%08X)", target_profile_->crc32);
+        ESP_LOGI(TAG, "MIDI Learn Wizard Completed Successfully! (CRC32: 0x%08X)", static_cast<unsigned int>(target_profile_->crc32));
         ESP_LOGI(TAG, "==================================================");
     } else if (current_step_ != LearnStep::Idle) {
         ESP_LOGI(TAG, ">>> STEP [%u/%u]: %s <<<", static_cast<unsigned>(current_step_), static_cast<unsigned>(LearnStep::Complete), currentStepName());
